pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imitation-learning-robotics"
version = "1.0.0"
description = "Modern imitation learning framework for robotics with behavioral cloning, DAgger, and GAIL"
authors = [{name = "AI Assistant", email = "ai@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.0.0",
    "imitation>=0.4.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "streamlit>=1.25.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

ros2 = [
    "rclpy>=3.3.0",
    "tf2-ros>=0.25.0",
    "geometry-msgs>=4.2.0",
    "sensor-msgs>=4.2.0",
    "nav-msgs>=4.2.0",
]

simulation = [
    "pybullet>=3.2.0",
    "mujoco>=2.3.0",
    "gymnasium-robotics>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/example/imitation-learning-robotics"
Repository = "https://github.com/example/imitation-learning-robotics"
Documentation = "https://github.com/example/imitation-learning-robotics#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
